Move PreprocessorCacheModeConfig to src/config.rs
