+++ src/cache/disk.rs
@@ -25,7 +25,8 @@ use std::time::{Duration, Instant};
 use crate::errors::*;
 
 use super::lazy_disk_cache::LazyDiskCache;
-use super::{PreprocessorCacheModeConfig, utils::normalize_key};
+use super::utils::normalize_key;
+use crate::config::PreprocessorCacheModeConfig;
 
 /// A cache that stores entries at local disk paths.
 pub struct DiskCache {
